backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3

# Video processing
yt-dlp>=2025.1.26
ffmpeg-python>=0.2.0

# AI/ML libraries
scenedetect[opencv]>=0.6.2
opencv-python-headless>=4.9.0.80
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.26.0
scipy>=1.12.0

# Face detection
mediapipe>=0.10.9

# Text detection (OCR)
easyocr>=1.7.1

# Job queue and caching
redis>=5.0.1
rq>=1.15.1

# Rate limiting
slowapi>=0.1.9

# File handling
aiofiles>=23.2.1

# Environment variables
python-dotenv>=1.0.0

# CORS
starlette>=0.35.1

# Scheduling and cleanup
apscheduler>=3.10.4

# Ollama for AI content analysis
ollama>=0.4.0
